refactor: move context menu to separate file
